+++ pkg/oauth2/oauth2.go
@@ -25,7 +25,7 @@ func GetController() (*Controller, error) {
 }
 
 // InitController use to init controller
-func InitController(ts token.Storage, us user.Storage, ds domain.Storage, as application.Storage, log logger.OpenAuthLogger, tokenType string, expiresIn int32) {
+func InitController(ts token.Storage, us user.Storage, ds domain.Storage, as application.Storage, log logger.OpenAuthLogger, tokenType string, expiresIn int64) {
 	once.Do(func() {
 		controller = &Controller{ts: ts, us: us, ds: ds, as: as, log: log, tokenType: tokenType, expiresIn: expiresIn}
 		controller.log.Debug("initial token controller successful")
@@ -41,5 +41,5 @@ type Controller struct {
 	as        application.Storage
 	log       logger.OpenAuthLogger
 	tokenType string
-	expiresIn int32
+	expiresIn int64
 }
